Added: Planarzone
Planarzone is the inner power layers of EasyEDA, and are processed as
negative layers, i.e. tracks drawn on this layer will be separators for
the real zones.

It should also be noted that EasyEDA uses an ill-formed path in
PLANARZONE objects, which is odd enough that we should put attention on
this.